refactor of ldap_comunication so it works with BerObject instead of char vectors

+++ src/BerSequenceObject.cpp
@@ -8,7 +8,7 @@ long long int BerSequenceObject::getLenght() {
 
   long long int dataLenght = 0;
 
-  for (int i = 0; i < objects.size(); i++) {
+  for (long long int i = 0; i < objects.size(); i++) {
     dataLenght += objects[i]->getLenght();
   }
 
@@ -20,13 +20,13 @@ std::vector<unsigned char> BerSequenceObject::getBerRepresentation() {
   int err = 0;
   std::vector<unsigned char> berRepresentation;
   long long int dataLenght = 0;
-  for (int i = 0; i < objects.size(); i++) {
+  for (long long int  i = 0; i < objects.size(); i++) {
     dataLenght += objects[i]->getLenght();
   }
 
   berRepresentation.push_back(tag);
   AppendLenght4Bytes(berRepresentation, dataLenght, &err);
-  for (int i = 0; i < objects.size(); i++) {
+  for (long long int  i = 0; i < objects.size(); i++) {
     std::vector<unsigned char> objectRepresentation =
         objects[i]->getBerRepresentation();
     berRepresentation.insert(berRepresentation.end(),
